backend/requirements.txt
Django==5.0
djangorestframework==3.15.1
django-filter==24.1
django-guardian==2.4.0
markdown==3.6
psycopg2==2.9.9